+++ src/library/DIPS.Mobile.UI/Components/Lists/ScrollView/ScrollView.cs
@@ -1,0 +1,6 @@
+namespace DIPS.Mobile.UI.Components.Lists;
+
+public partial class ScrollView : Microsoft.Maui.Controls.ScrollView
+{
+    
+}